requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
cryptography==42.0.8
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.6
django-celery-beat==2.6.0
django-debug-toolbar==4.2.0
django-filter==23.4
django-import-export==3.3.3
django-redis==5.4.0
django-simple-history==3.4.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
flower==2.0.1
humanize==4.9.0
idna==3.7
kombu==5.3.7
MarkupPy==1.14
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.1.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
pycparser==2.22
PyJWT==2.8.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.6
requests==2.32.3
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.4.4
tablib==3.5.0
tornado==6.4.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
xlrd==2.0.1
xlwt==1.3.0
pylint==3.2.3
pylint-django==2.5.5
drf-yasg==1.21.7
django-allauth==0.63.3